Transpiler: INTEGER inference — index-shaped numerics emit as C# int (W0026)
decimal stays the default for Harbour numerics, but a LOCAL/STATIC
whose numeric life is purely index-shaped — array subscripts, hash
keys, FOR variables, provably-integral initializers and arithmetic
(int literals, Len()/At()/RAt()/Asc()/int(), INTEGER-returning user
functions, self-references) — upgrades NUMERIC → INTEGER in a new
Pass 2.5 and emits as C# int. int widens to decimal implicitly, so
reads cost nothing anywhere; subscripts drop the (int) cast.

Disqualifiers, per Alex's no-silent-demotion rule:
* hard (division operand or division/fractional-literal assignment —
  Harbour n/2 is 1.5, C# int/int truncates): stays decimal and, when
  index-used, fires W0026 at the disqualifying site (wrap with int()
  or keep decimal);
* soft (non-integral assignment RHS, @by-ref pass, SWITCH subject —
  case labels are const decimal defines): stays decimal, reported
  only by --type-audit as INT-CONFLICT.

Writes into int-typed storage get an (int) coercion at five sites:
LOCAL init, assignment, FOR start, and source-declared
class members' INIT and Self-assignments — the dialog-model family's
 declarations never compiled before (their define-constant
inits and decimal-param assignments were CS0266 since 2026-07-04).

Cross-function: INTEGER returns flow through the reftab; INTEGER vs
NUMERIC merges quietly to NUMERIC in both return collection and
call-site param refinement (a refinement, never a W0022 conflict).
genhb degrades INTEGER to NUMERIC (the lexer has no AS INTEGER).

Audit refinements: the TSV now opens with a self-describing header
for every category; INT-CANDIDATE (the pre-implementation
measurement, 959 clean candidates) is retired — clean candidates
now simply ARE int, and the report stays insufficiency-only.

test78 covers loops, index vars, Len() inits, integral arithmetic,
the FirstReal cross-function chain, W0026 on a division-fed index,
and int-into-decimal widening, across prg/hb/cs.

easipos: 597 → 586 errors (fixes outweigh the newly-visible inline-
method subscript cluster, CS0266 ×6, which was always semantically
broken — the textual INLINE translator does no 1-based conversion —
and is now loudly broken instead). Scan stays clean: zero W0026s
across 670 files.

+++ src/transpiler/tests/csout/test13.cs
@@ -23,7 +23,7 @@ public static partial class Program
     public static decimal SumEvenTo(decimal nMax = default)
     {
         decimal nSum = 0;
-        decimal i = default;
+        int i = default;
 
         for (i = 1; i <= nMax; i++)
         {
@@ -45,12 +45,12 @@ public static partial class Program
     {
         string cItem = default;
         decimal nPos = 0;
-        decimal i = default;
+        int i = default;
 
         // Nested FOR with EXIT
         for (i = 1; i <= HbRuntime.Len(aItems); i++)
         {
-            if (aItems[(int)(i) - 1] == cTarget)
+            if (aItems[i - 1] == cTarget)
             {
                 nPos = i;
                 HbRuntime.QOut("nPos=" + HbRuntime.Str(nPos));
