+++ src/transpiler/tests/csout/test30.cs
@@ -29,7 +29,7 @@ public static partial class Program
 {
     public static void Main(string[] args)
     {
-        decimal i = default;
+        int i = default;
         decimal nHits = 0;
 
         for (i = 1; i <= 10; i++)
