+++ src/transpiler/tests/csout/test5.cs
@@ -7,7 +7,7 @@ public static partial class Program
 {
     public static void Main(string[] args)
     {
-        decimal i = default;
+        int i = default;
         decimal nSum = 0;
         dynamic[] aItems = new dynamic[] { "apple", "banana", "cherry" };
         string cItem = default;
