+++ src/transpiler/tests/csout/test60.cs
@@ -18,7 +18,7 @@ public static partial class Program
         Dictionary<string, dynamic> hData = new Dictionary<string, dynamic> { { "a", 1 }, { "b", 2 }, { "c", 3 } };
         // note the duplicate 1
         dynamic[] aNums = new dynamic[] { 3, 1, 2, 1 };
-        decimal i = default;
+        int i = default;
         string cOut = default;
 
         HbRuntime.QOut("round_2.5=" + HbRuntime.LTrim(HbRuntime.Str(HbRuntime.Round(2.5m, 0))));
@@ -31,7 +31,7 @@ public static partial class Program
         cOut = "";
         for (i = 1; i <= HbRuntime.Len(aNums); i++)
         {
-            cOut += HbRuntime.LTrim(HbRuntime.Str(aNums[(int)(i) - 1]));
+            cOut += HbRuntime.LTrim(HbRuntime.Str(aNums[i - 1]));
         }
 
         HbRuntime.QOut("sorted=" + cOut);
